+++ pkg/webui/console/components/payload-formatters-form/index.js
@@ -38,7 +38,10 @@ const m = defineMessages({
   repository: 'Repository',
   formatterType: 'Formatter type',
   formatterParameter: 'Formatter parameter',
-  grpcDescription: 'The address of the service to connect to',
+  grpcFieldDescription: 'The address of the service to connect to',
+  appFormatter: 'Use application payload formatter',
+  appFormatterWarning:
+    'This option sets both uplink and downlink formatters to application link defaults',
 })
 
 const FIELD_NAMES = {
@@ -80,7 +83,7 @@ class PayloadFormattersForm extends React.Component {
 
   @bind
   async handleSubmit(values, { resetForm }) {
-    const { onSubmit, onSubmitSuccess, onSubmitFailure } = this.props
+    const { onSubmit, onSubmitSuccess, onSubmitFailure, uplink } = this.props
 
     this.setState({ error: '' })
 
@@ -94,18 +97,18 @@ class PayloadFormattersForm extends React.Component {
       [FIELD_NAMES.RADIO]: type,
     }
 
-    let parameter = ''
+    let parameter
     switch (type) {
       case TYPES.JAVASCRIPT:
         parameter = javascriptParameter
         resetValues[FIELD_NAMES.JAVASCRIPT] = javascriptParameter
+        resetValues[FIELD_NAMES.GRPC] = getDefaultGrpcServiceFormatter(uplink)
         break
       case TYPES.GRPC:
         parameter = grpcParameter
+        resetValues[FIELD_NAMES.JAVASCRIPT] = getDefaultJavascriptFormatter(uplink)
         resetValues[FIELD_NAMES.GRPC] = grpcParameter
         break
-      default:
-        parameter = undefined
     }
 
     try {
@@ -153,7 +156,7 @@ class PayloadFormattersForm extends React.Component {
             name={FIELD_NAMES.GRPC}
             type="text"
             placeholder={sharedMessages.addressPlaceholder}
-            description={m.grpcDescription}
+            description={m.grpcFieldDescription}
           />
         )
       default:
@@ -162,11 +165,11 @@ class PayloadFormattersForm extends React.Component {
   }
 
   render() {
-    const { initialType, initialParameter, linked, uplink } = this.props
-    const { error } = this.state
+    const { initialType, initialParameter, linked, uplink, allowReset } = this.props
+    const { error, type } = this.state
 
     const initialValues = {
-      [FIELD_NAMES.RADIO]: initialType,
+      [FIELD_NAMES.RADIO]: type,
       [FIELD_NAMES.JAVASCRIPT]:
         initialType === TYPES.JAVASCRIPT ? initialParameter : getDefaultJavascriptFormatter(uplink),
       [FIELD_NAMES.GRPC]:
@@ -189,7 +192,9 @@ class PayloadFormattersForm extends React.Component {
             title={m.formatterType}
             component={Radio.Group}
             onChange={this.onTypeChange}
+            warning={type === TYPES.DEFAULT ? m.appFormatterWarning : undefined}
           >
+            {allowReset && <Radio label={m.appFormatter} value={TYPES.DEFAULT} />}
             <Radio label={sharedMessages.none} value={TYPES.NONE} />
             <Radio label="Javascript" value={TYPES.JAVASCRIPT} />
             <Radio label={m.grpc} value={TYPES.GRPC} />
@@ -207,6 +212,7 @@ class PayloadFormattersForm extends React.Component {
 }
 
 PayloadFormattersForm.propTypes = {
+  allowReset: PropTypes.bool,
   initialParameter: PropTypes.string,
   initialType: PropTypes.oneOf(Object.values(TYPES)).isRequired,
   linked: PropTypes.bool.isRequired,
@@ -220,6 +226,7 @@ PayloadFormattersForm.defaultProps = {
   initialParameter: '',
   onSubmitSuccess: () => null,
   onSubmitFailure: () => null,
+  allowReset: false,
 }
 
 export default PayloadFormattersForm